modify for windows compatible

+++ ofkMatrixHelper.h
@@ -12,7 +12,7 @@ public:
     //Caution, this function will enable GL_DEPTH_TEST, and clear GL_DEPTH_BUFFER_BIT.
     //and in this function, viewport, projection Matrix and modelView matrix will be changed.
     
-    static ofVec3f getUnProjectionPoint(float ofScreenPosX, float ofScreenPosY, const float *model, const float *proj, const GLint *view);
+    static ofVec3f getUnProjectionPoint(float ofScreenPosX, float ofScreenPosY, const GLdouble *model, const GLdouble *proj, const GLint *view);
     static ofVec3f getUnProjectionPoint(float ofScreenPosX, float ofScreenPosY, ofMatrix4x4 modelView, ofMatrix4x4 projection, float viewWidth, float viewHeight );
 	static ofVec3f getProjectionPoint(ofVec3f pos, const GLdouble *model, const GLdouble *proj, const GLint *view);
